Extract composer order diff UI boundary
