+++ assets/js/composer-post-commit-state.js
@@ -1,4 +1,4 @@
-import { parseEncryptedMarkdownEnvelope } from './encrypted-content.js?v=press-system-v3.4.23';
+import { parseEncryptedMarkdownEnvelope } from './encrypted-content.js?v=press-system-v3.4.24';
 
 export function createPostCommitStateApplier({
   stagingRegistry,
