+++ assets/js/hieditor.js
@@ -1,4 +1,4 @@
-import { simpleHighlight } from './syntax-highlight.js?v=press-system-v3.4.23';
+import { simpleHighlight } from './syntax-highlight.js?v=press-system-v3.4.24';
 
 function escapeHtmlInline(text) {
   if (!text) return '';
